+++ src/palettes/DarkPlusHighContrastPalette.ts
@@ -5,4 +5,6 @@ export const DarkPlusHighContrastPalette: ThemeColorPalette = {
     ...DarkPlusPalette,
     Foreground: "#ffffff",
     Background: "#000000",
+    Primary:    "#f38518",
+    Secondary:  "#9cdcfe",
 }