+++ src/palettes/LightPlusPalette.ts
@@ -2,7 +2,28 @@ import { ThemeColorPalette } from "@types"
 import { DarkPlusPalette } from "."
 
 export const LightPlusPalette: ThemeColorPalette = {
-    ...DarkPlusPalette,
-    Foreground: "#000000",
-    Background: "#ffffff",
+    // Semantic
+    RegExp:      "#811F3F",
+    Strings:     "#A31515",
+    Chars:       "#EE0000",
+    Types:       "#267F99",
+    Variables:   "#0070C1",
+    Functions:   "#795E26",
+    Numbers:     "#098658",
+    Keywords:    "#0000FF",
+    FlowControl: "#AF00DB",
+    Comments:    "#008000",
+    DocComments: "#008000",
+    Tags:        "#1e1e1e",
+    Noise:       "#a4a4a4",
+    // UI
+    Foreground:  "#1e1e1e",
+    Background:  "#d4d4d4",
+    Selection:   "#ADD6FF80",
+    Primary:     "#008000",
+    Secondary:   "#AF00DB",
+    Error:       "#811F3F",
+    Warning:     "#795E26",
+    Info:        "#0070C1",
+    Success:     "#008000",
 }